+++ time-macros/src/format_description/modifier.rs
@@ -74,6 +74,7 @@ to_tokens! {
     pub(crate) struct Month {
         pub(crate) padding: Padding,
         pub(crate) repr: MonthRepr,
+        pub(crate) case_sensitive: bool,
     }
 }
 
@@ -96,6 +97,7 @@ to_tokens! {
     pub(crate) struct Weekday {
         pub(crate) repr: WeekdayRepr,
         pub(crate) one_indexed: bool,
+        pub(crate) case_sensitive: bool,
     }
 }
 
@@ -146,6 +148,7 @@ to_tokens! {
 to_tokens! {
     pub(crate) struct Period {
         pub(crate) is_uppercase: bool,
+        pub(crate) case_sensitive: bool,
     }
 }
 
@@ -219,12 +222,14 @@ impl_default! {
     Month => Self {
         padding: Padding::default(),
         repr: MonthRepr::default(),
+        case_sensitive: true,
     };
     Ordinal => Self { padding: Padding::default() };
     WeekdayRepr => Self::Long;
     Weekday => Self {
         repr: WeekdayRepr::default(),
         one_indexed: true,
+        case_sensitive: true,
     };
     WeekNumberRepr => Self::Iso;
     WeekNumber => Self {
@@ -243,7 +248,7 @@ impl_default! {
         is_12_hour_clock: false,
     };
     Minute => Self { padding: Padding::default() };
-    Period => Self { is_uppercase: true };
+    Period => Self { is_uppercase: true, case_sensitive: true };
     Second => Self { padding: Padding::default() };
     SubsecondDigits => Self::OneOrMore;
     Subsecond => Self { digits: SubsecondDigits::default() };
@@ -269,6 +274,7 @@ pub(crate) struct Modifiers {
     pub(crate) year_repr: Option<YearRepr>,
     pub(crate) year_is_iso_week_based: Option<bool>,
     pub(crate) sign_is_mandatory: Option<bool>,
+    pub(crate) case_sensitive: Option<bool>,
 }
 
 impl Modifiers {
@@ -334,6 +340,12 @@ impl Modifiers {
                 | (b"year", b"padding:none") => modifiers.padding = Some(Padding::None),
                 (b"hour", b"repr:24") => modifiers.hour_is_12_hour_clock = Some(false),
                 (b"hour", b"repr:12") => modifiers.hour_is_12_hour_clock = Some(true),
+                (b"month", b"case_sensitive:true")
+                | (b"period", b"case_sensitive:true")
+                | (b"weekday", b"case_sensitive:true") => modifiers.case_sensitive = Some(true),
+                (b"month", b"case_sensitive:false")
+                | (b"period", b"case_sensitive:false")
+                | (b"weekday", b"case_sensitive:false") => modifiers.case_sensitive = Some(false),
                 (b"month", b"repr:numerical") => modifiers.month_repr = Some(MonthRepr::Numerical),
                 (b"month", b"repr:long") => modifiers.month_repr = Some(MonthRepr::Long),
                 (b"month", b"repr:short") => modifiers.month_repr = Some(MonthRepr::Short),